requirements.txt
torch
mmcv-full
mmdet>=2.24.1
mmrotate>=0.2.0
numpy
matplotlib
seaborn
setuptools
terminaltables
opencv-python